#39 - feat: add Socket.IO for fetching new notifications
This commit adds Socket.IO to the React app to listen for new
notifications from the server and fetch new notifications when the user
is logged in.

The component PrimaryHeader subscribes to new notifications and displays
the list of notifications in real-time. Update useRefreshHook and logout
requests to reset user notifications when their session expired.
Notifications will be fetched when the user login, refreshes the new
token success or the server emits a signal to them.

This feature improves the user experience of the app by providing
real-time notifications without the need for manual refreshing.