#43 - fix: resolve issue with navigation to Notification page
The app was crashing when navigating to the Notification page. This was
due to the 'option selected' variable being set by default "account
profile" the first time the user login to the website. So that
notifications["account profile"] is undefined. What we expected is
notifications["news"].

I added a statement to update the sidebar selection to "news" before
navigating to the notifications page and tested the app to confirm that
the issue is now resolved.

Fixes #43