#48 - fix: layout reflow cause by lazyload img
Previously, the lazy loading of product images on our website was
causing a layout reflow issue, leading to a poor user experience.
This commit fixes the bug by implementing a solution that uses a fixed
the aspect ratio for the image container to prevent layout shifting and
compressing the reserved image for loading faster.

As a result, users can now enjoy a smoother browsing experience on our
a website with no unexpected layout shifts caused by the lazy-loading of
product images.